Make sure that returned callback url as field value always valid and move common parts from developer app to app entity.
